pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bhre-project"
version = "0.1.0"
authors = [{name = "Your Name", email = "you@example.com"}]
description = "Bayesian Hypervector Rule Embeddings: Online, explainable, no-backprop hypervector library"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "torch>=2.0",
  "numpy>=1.23",
  "pyyaml>=6.0"
]

[project.scripts]
bhre = "amt.cli:main"

[project.optional-dependencies]
dev = ["pytest", "mypy", "flake8", "build"]
